chore: remove legacy "doodle" logo-variant feature
The `logos` table backed special-occasion logo variants: on a matching
calendar day a random active logo was served in place of the normal
Freegle logo, much like a Google Doodle. The feature is being retired.

Server (V2 Go API):
- Delete the logo package and the GET /logo endpoint (route, swagger
  annotation, generated swagger.json entry).
- Preserve the auth-middleware stale-JWT regression tests, which only
  used /logo as a convenient public endpoint, by moving them to
  test/authmiddleware_test.go and re-pointing the public case at /online.

Client (nuxt3):
- Delete the logo Pinia store, LogoAPI, and their registration/init in
  app.vue, modtools/app.vue and api/index.js / api/index.d.ts.
- Remove the doodle fetch/override from useNavbar; NavbarDesktop now
  renders the standard logo unconditionally (dead gif branch removed).
- Trim the affected unit tests and drop stores/logo.spec.js.

Database:
- Add migration 2026_07_09_000001_drop_logos_table (down recreates it).
- Include paired idempotent production SQL to drop the table.

V1 (iznik-server) intentionally untouched; it has already been removed.
