--- iznik-nuxt3/api/LogoAPI.js
@@ -1,7 +1,0 @@
-import BaseAPI from '@/api/BaseAPI'
-
-export default class LogoAPI extends BaseAPI {
-  async fetch(params) {
-    return await this.$getv2('/logo', params)
-  }
-}
