+++ iznik-nuxt3/app.vue
@@ -137,7 +137,6 @@ import { useStatsStore } from './stores/stats'
 import { useMicroVolunteeringStore } from './stores/microvolunteering'
 import { useImageStore } from './stores/image'
 import { useDomainStore } from './stores/domain'
-import { useLogoStore } from './stores/logo'
 import { useLocationStore } from './stores/location'
 import { useShortlinkStore } from './stores/shortlinks'
 import { useMiscStore } from './stores/misc'
@@ -230,7 +229,6 @@ const statsStore = useStatsStore()
 const microVolunteeringStore = useMicroVolunteeringStore()
 const imageStore = useImageStore()
 const domainStore = useDomainStore()
-const logoStore = useLogoStore()
 const locationStore = useLocationStore()
 const shortlinkStore = useShortlinkStore()
 
@@ -263,7 +261,6 @@ statsStore.init(runtimeConfig)
 microVolunteeringStore.init(runtimeConfig)
 imageStore.init(runtimeConfig)
 domainStore.init(runtimeConfig)
-logoStore.init(runtimeConfig)
 locationStore.init(runtimeConfig)
 shortlinkStore.init(runtimeConfig)
