+++ iznik-nuxt3/tests/unit/components/NavbarDesktop.spec.js
@@ -6,8 +6,6 @@ import NavbarDesktop from '~/components/NavbarDesktop.vue'
 const mockNavbar = {
   online: ref(true),
   distance: ref(5),
-  logo: ref('/logo.png'),
-  logoFormat: ref('png'),
   unreadNotificationCount: ref(3),
   chatCount: ref(2),
   activePostsCount: ref(5),
@@ -66,10 +64,6 @@ describe('NavbarDesktop', () => {
             props: ['variant', 'disabled'],
             emits: ['click'],
           },
-          'b-img': {
-            template: '<img class="b-img" :src="src" :alt="alt" />',
-            props: ['src', 'format', 'alt'],
-          },
           'b-badge': {
             template: '<span class="b-badge" :class="variant"><slot /></span>',
             props: ['variant', 'title'],
@@ -119,12 +113,6 @@ describe('NavbarDesktop', () => {
       mockNavbar.online.value = false
       const wrapper = createWrapper()
       expect(wrapper.find('.offline-indicator').exists()).toBe(true)
-    })
-
-    it('shows gif logo directly when format is gif', () => {
-      mockNavbar.logoFormat.value = 'gif'
-      const wrapper = createWrapper()
-      expect(wrapper.find('.b-img').exists()).toBe(true)
     })
   })
