+++ iznik-server-go/swagger/swagger.go
@@ -1990,19 +1990,6 @@ type locationsResponse struct {
 //	401: errorResponse
 
 // ============================================================================
-// Logo
-// ============================================================================
-
-// swagger:route GET /logo misc getLogo
-// Get logo
-//
-// Returns logo information
-//
-// Responses:
-//
-//	200: genericResponse
-
-// ============================================================================
 // Membership
 // ============================================================================
